Fixed (sort of) the bottom layout and node-text visibility.
Changes:

Added KibitzNodeText as its own component.
Draft mode now separates the bottom stack into:
analyze toolbar
editable move comment box
variation name/post controls
move tree
Posted variation viewing now shows the move comment box read-only, so users can read per-node text while watching.
Updated desktop equal-pane CSS so the textarea no longer sits in a too-short row and overlaps the toolbar/tree.
Added the same node-text visibility to mobile Compare.
